support material theme

+++ app/src/main/java/com/temple/skiaui/compose/ui/util/Colors_androidKt.kt
@@ -3,5 +3,5 @@ package com.temple.skiaui.compose.ui.util
 import androidx.compose.ui.graphics.Color
 
 internal fun composeColorToSkiaColor(color: Color): Int {
-    return (color.value shr 32).toUInt().toInt()
+    return (color.value shr 32).toInt()
 }
